utils/test_minimal_system.py: Small fixes
- Do not hardcode x86_64 unnecessarily.
- Use shlex to split commandline args.
- Use python standard logging.
- Ensure we don't overwrite the dialog lines.

+++ utils/test_minimal_system.py
@@ -21,11 +21,12 @@ import asyncio
 import asyncio.subprocess
 import logging
 import os
+import shlex
 import signal
 import subprocess
 import sys
 
-QEMU = "qemu-system-x86_64"
+QEMU = f"qemu-system-{os.uname().machine}"
 QEMU_EXTRA_ARGS = ["-m", "256"]
 
 FAILURE_TIMEOUT = 300  # seconds
@@ -78,8 +79,8 @@ def build_qemu_image_command(args):
         out = subprocess.check_output([QEMU, "-accel", "help"], encoding="ascii")
         if "kvm" in out.splitlines():
             kvm_args = ["-enable-kvm"]
-    except subprocess.CalledProcessError:
-        sys.stderr.write("Cannot query qemu for accelerator. Not using it.\n")
+    except subprocess.CalledProcessError as e:
+        logger.debug("Cannot query qemu for accelerator. Not using it: %s", e)
 
     return (
         [QEMU, "-drive", f"file={args.sda},format=raw", "-nographic"]
@@ -89,7 +90,7 @@ def build_qemu_image_command(args):
 
 
 def build_command(args):
-    return args.command.split()
+    return shlex.split(args.command)
 
 
 def argument_parser():
@@ -116,7 +117,7 @@ def argument_parser():
 
 
 async def await_line(stream, marker):
-    """Read from 'stream' until a line appears contains 'marker'."""
+    """Read from 'stream' until a line contains 'marker'."""
     marker = marker.encode("utf-8")
     buf = b""
 
@@ -127,20 +128,19 @@ async def await_line(stream, marker):
         lines = buf.split(b"\n")
         for line in lines:
             if marker in line:
-                try:
-                    return line.decode("utf-8")
-                except UnicodeDecodeError:
-                    break
+                return line.decode("utf-8", errors="replace")
         buf = lines[-1]
+    return None
 
 
 async def run_test(command, dialog):
-    dialog = DIALOGS[dialog]
+    dialog = DIALOGS[dialog].copy()
 
     success = False
+    process = None
 
     try:
-        logger.debug("Starting process: %s", command)
+        logger.info("Starting process: %s", command)
         process = await asyncio.create_subprocess_exec(
             *command,
             stdin=asyncio.subprocess.PIPE,
@@ -154,17 +154,19 @@ async def run_test(command, dialog):
             assert prompt is not None
             if dialog:
                 process.stdin.write(dialog.pop(0).encode("ascii") + b"\n")
+                await process.stdin.drain()
 
-        print("Test successful")
+        process.stdin.close()
+        logger.info("Test successful")
         success = True
     finally:
-        try:
-            os.killpg(os.getpgid(process.pid), signal.SIGKILL)
-        except ProcessLookupError:
-            pass
+        if process is not None:
+            try:
+                os.killpg(os.getpgid(process.pid), signal.SIGKILL)
+            except ProcessLookupError:
+                pass
 
-        await process.communicate()
-        await process.wait()
+            await process.communicate()
 
     return success
 
@@ -179,7 +181,7 @@ def main():
     try:
         result = asyncio.run(task)
     except asyncio.TimeoutError:
-        print("VM was considered inresponsive and test was aborted", file=sys.stderr)
+        logger.info("VM was considered unresponsive and test was aborted")
         return 1
 
     if result:
